Update README
The --debug flag is deprecated in my current version of Node (v8.11.2) which seems to be more commonly used than past versions. Also, changed the port number to 3000 since it's also a more common port for running servers than 9222—hoping to make people more familiar and comfortable with what they see in the docs.